refactor: delete Button component from Keypad
