+++ src/components/Keypad/styled.js
@@ -1,6 +1,7 @@
+import { DRAW_HISTORY_BUTTONS } from '@constants/options';
 import styled, { css } from 'styled-components';
 
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
     grid-area: keypad;
     display: grid;
     grid-template-areas:
@@ -23,4 +24,29 @@ const Wrapper = styled.div`
     `};
 `;
 
-export default Wrapper;
+export const Button = styled.button`
+    ${({
+        area,
+        theme: {
+            fontSize,
+            colors,
+            background,
+            border: { original },
+        },
+    }) => css`
+        grid-area: ${area};
+        color: ${colors.black};
+        border: ${original};
+        background: ${DRAW_HISTORY_BUTTONS.includes(area)
+            ? background.special
+            : ''};
+        font-size: ${fontSize.main}px;
+    `};
+    cursor: pointer;
+    &:focus-visible {
+        outline: none;
+    }
+    &:active {
+        background-color: white;
+    }
+`;
